Refactor module structure and improve code organization
