+++ blender/src/lib.rs
@@ -1,7 +1,4 @@
-// TODO: Add documentation and examples to this code.
 pub mod blender;
-// #[cfg(feature = "manager")]
 pub mod manager;
 pub mod models;
-// #[cfg(feature = "manager")]
 pub mod page_cache;
