+++ blender/src/models.rs
@@ -1,10 +1,8 @@
 pub mod args;
 pub mod blender_peek_response;
 pub mod blender_render_setting;
-pub mod device;
-
-// #[cfg(feature = "manager")]
 pub mod category;
+pub mod device;
 pub mod download_link;
 pub mod engine;
 pub mod format;
