+++ src-tauri/Cargo.toml
@@ -17,14 +17,12 @@ crate-type = ["staticlib", "cdylib", "rlib"]
 tauri-build = { version = "^2.0.1", features = [] }
 
 [dependencies]
-bincode = "1.3.3"
-anyhow = "^1.0.89"
+anyhow = "^1.0"
 async-trait = "0.1.83"
 async-std = "1.13.0"
 blend = "0.8.0"
 blender = { path = "./../blender/" }
-dirs = "5.0.1"
-libp2p = { version = "^0", features = [
+libp2p = { version = "^0.54", features = [
     "mdns",
     "macros",
     "gossipsub",
@@ -38,29 +36,31 @@ libp2p = { version = "^0", features = [
     "kad",
     "request-response",
 ] }
-pnet = "0.34"
-semver = "^1.0.22"
-tauri-cli = "2.0.3"
+bincode = "1.3.3"
+dirs = "5.0.1"
+pnet = "^0.35"
+semver = "^1.0"
+thiserror = "*"
+tauri-cli = "^2.1"
 tauri-plugin-core = "2.0.0-beta.0"
-tauri-plugin-dialog = "^2"
-tauri-plugin-fs = "^2"
-tauri-plugin-os = "^2"
-tauri-plugin-persisted-scope = "^2"
-tauri-plugin-shell = "^2"
+tauri-plugin-dialog = "^2.0"
+tauri-plugin-fs = "^2.0"
+tauri-plugin-os = "^2.0"
+tauri-plugin-persisted-scope = "^2.0"
+tauri-plugin-shell = "^2.0"
 tokio = { version = "^1.41", features = ["full"] }
-thiserror = "^1"
-quinn = "0.11.6"
-tracing = "0.1.40"
-tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
+quinn = "^0.11"
+tracing = "^0.1"
+tracing-subscriber = { version = "^0.3", features = ["env-filter"] }
 clap = "4.5.21"
 
 # this came autogenerated. I don't think I will develop this in the future, but would consider this as an april fools joke. Yes I totally would.
 [target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
-tauri-plugin-cli = "2"
-tauri = { version = "^2.0.0", features = ["protocol-asset"] }
-serde = { version = "^1.0.210", features = ["derive"] }
-serde_json = "^1.0.128"
-uuid = { version = "^1.10.0", features = [
+tauri-plugin-cli = "^2"
+tauri = { version = "^2.0", features = ["protocol-asset"] }
+serde = { version = "^1.0", features = ["derive"] }
+serde_json = "^1.0"
+uuid = { version = "^1.10", features = [
     "v4",
     "fast-rng",
     "macro-diagnostics",
