+++ src-tauri/src/models/mod.rs
@@ -1,7 +1,10 @@
 pub mod app_state;
+pub mod behaviour;
 pub(crate) mod common;
 pub mod error;
 pub(crate) mod job;
+pub mod message;
+pub mod network;
 pub(crate) mod project_file;
 pub(crate) mod render_info;
 pub mod render_queue;
